Revision 4
Bullet Physics integrated into the engine.
Gameobjects rescaled to 1% of their previous size to accommodate for the
scaling of Bullet Physics.

+++ Alchemist/Alchemist/Camera.h
@@ -70,5 +70,5 @@ private:
 	float farPlane;
 	float fieldOfView;
 	float aspectRatio;
-	int wndWidth, wndHeight;
+	float wndWidth, wndHeight;
 };
